fix(sentry-infra-tools): Use the dependencies already present on sentry-infra-tools
Since I am porting over the sentry-infra-tools from the ops repo, I would
like to ensure that I do not inadvertently break the tooling. In order to do
that I would like to use the same dependencies which exist in the ops
repo instead of using their latest versions.

Once we verify everything works correctly, we can think of upgrading the
package versions.