Add `tokenizers` package (#1675)
`tokenizers` is used for tokenization of strings. While we do currently import titoken as well, we want to use `tokenizers` to
count tokens the same way Seer does in Sentry for stacktrace token counts. We could consider converting usages of titoken
to `tokenizers` to maintain a dependency on only one tokenization package, but for now we will have them both.